refactor: extract time functionality to time_util.py
